--- a/gemrb/core/ArchiveImporter.h
+++ b/gemrb/core/ArchiveImporter.h
@@ -35,6 +35,7 @@ public:
 	//decompressing a .sav file similar to CBF
 	virtual int DecompressSaveGame(DataStream *compressed) = 0;
 	virtual int AddToSaveGame(DataStream *str, DataStream *uncompressed) = 0;
+	virtual int AddToSaveGameCompressed(DataStream *str, DataStream *compressed) = 0;
 };
 
 }