--- a/gemrb/core/ArchiveImporter.h
+++ b/gemrb/core/ArchiveImporter.h
@@ -24,6 +24,7 @@
 #include "globals.h"
 
 #include "Plugin.h"
+#include "SaveGameAREExtractor.h"
 
 namespace GemRB {
 
@@ -33,7 +34,7 @@ public:
 	~ArchiveImporter(void) override;
 	virtual int CreateArchive(DataStream *stream) = 0;
 	//decompressing a .sav file similar to CBF
-	virtual int DecompressSaveGame(DataStream *compressed) = 0;
+	virtual int DecompressSaveGame(DataStream *compressed, SaveGameAREExtractor&) = 0;
 	virtual int AddToSaveGame(DataStream *str, DataStream *uncompressed) = 0;
 	virtual int AddToSaveGameCompressed(DataStream *str, DataStream *compressed) = 0;
 };