--- a/gemrb/core/CMakeLists.txt
+++ b/gemrb/core/CMakeLists.txt
@@ -61,6 +61,7 @@ FILE(GLOB gemrb_core_LIB_SRCS
 	Resource.cpp
 	ResourceDesc.cpp
 	ResourceManager.cpp
+	SaveGameAREExtractor.cpp
 	SaveGameIterator.cpp
 	SaveGameMgr.cpp
 	ScriptEngine.cpp