## Description
I noticed that save games of later stages in the game take notably longer: Every ARE is extracted into the cache onto the disk. Yet, we don't need all these ARE files right now, and given the progress, some of them never again.

BG1 spills around 260 ARE files into the cache folder, other games a little less for me.

Why not extract only when we really need it? On my very fast NVMe this _only_ saves half a second (and a lot of console entries), but surely some more on HDDs or RPIs -- unfortunately no test hardware at hand right now.

Feel free to discuss first, but I considered this an avoidable practice.

## Checklist

- [x] Commit messages are descriptive and explain the rationale for changes
- [x] I used the same coding style as the surrounding code <!-- yet to be formally defined, see issue #161 -->
- [x] I have tested the proposed changes
- [x] I extended the documentation, if necessary
- [x] The proposed change builds also on our build bots (check after submission)