@MarcelHB 2021-07-04
Let me check one gap tomorrow again: When we repack the SAV file by saving and then try to enter some older area, the look up cache probably needs to be rebuild.