@lynxlynxlynx 2021-07-04
Interesting! 
Game::LoadMap gets called also from other (nonsave) places iirc - is that part fine?