@MarcelHB 2021-07-04
That's not just fine but a requirement, since we don't really want to care who requests a map change, only that we may need to emit it from the running savegame.