@MarcelHB 2021-07-04
Now, we have the following status:
* Loading is roughly constant.
* The first save gets the penalty of about what we saved in the loading, since every sleeping ARE is extracted before repacking. Any subsequent save is spared, but of course everything will be packed again, and again.

Room for even more improvement: Why extracting, and repacking? It involves CPU and filesystem for no reason. Instead, we could copy the remaining, compressed entries over into a new savegame before picking up the cache stuff. With some precaution of course, if we update the same savegame we are currently feeding from.