@MarcelHB 2021-07-04
I'd like to invite you for some testing. Here is a BG save game to do some stress-testing if useful: 
[000000003-1.zip](https://github.com/gemrb/gemrb/files/6760737/000000003-1.zip)

So what's going on now:

- Extract ARE files only on demand from the save game that is know to have loaded the current session. Keep the rest in a registry that knows what's still in the box.
- When saving into a different save game: raw-copy the compressed data into the new save game. So no more decompressing, FS bothering and recompression. Continue with the ARE files from the cache folder as before.
- When saving into the same save game: drop a `ares.blb` into the cache folder (that is sequence of all unextracted AREs) and pick this up into the new SAV file. The registry is getting updated by the new locations, so we can safely change maps after the old SAV is no more. Continue with the ARE files from the cache folder as before. 

The results are quite promising to me.