@lynxlynxlynx 2021-07-05
I think it would be worthwhile to add some timing logging, so we can test the benefits more easily. I'd just log the total duration once the process is over.

The ees strongly improved the save times, as they're practically instant there.