@MarcelHB 2021-07-05
What do you want to see exactly? I can emit different pieces of ms spent for the steps in saving and loading (which need to be summed then) but to see the difference, you'd need to do some different logging in master as well.

Saving is now virtually instant here, too.