@lynxlynxlynx 2021-07-05
Ideally it'd be a separate pr, that we can merge first, gather results without the patches and then also with them.

As for data, for me the totals would be enough. Like maybe the time spent in LoadGame.