@bradallred 2021-07-05
We can just use perf or other profiling tools. Not really a need to add our own instrumentation. 

> On Jul 5, 2021, at 10:31, Jaka Kranjc ***@***.***> wrote:
> 
> ﻿
> Ideally it'd be a separate pr, that we can merge first, gather results without the patches and then also with them.
> 
> As for data, for me the totals would be enough. Like maybe the time spent in LoadGame.
> 
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.