@MarcelHB 2021-07-05
So, benchmarks. I put my `Cache2` onto some old USB 2.0 FAT32 key to simulate something from real life, but slower than my NVMe. This is for amplifying the effect.

Loading the BG save game from above:
master: ~8.5s (!!)
this: ~900ms

Saving the same:
master: ~900ms
this: ~80ms

For an NVMe, loading:
master: ~900ms
this: ~300ms

Saving: neglectable

I've put a temporary commit on top of this branch, and if you like to see yourself: [for master](https://github.com/gemrb/gemrb/compare/master...MarcelHB:are_loading_old?expand=1.)

> 
> We can just use perf or other profiling tools. Not really a need to add our own instrumentation.
> […](#)
> On Jul 5, 2021, at 10:31, Jaka Kranjc ***@***.***> wrote: ﻿ Ideally it'd be a separate pr, that we can merge first, gather results without the patches and then also with them. As for data, for me the totals would be enough. Like maybe the time spent in LoadGame. — You are receiving this because you are subscribed to this thread. Reply to this email directly, view it on GitHub, or unsubscribe.

You are right, but I don't know how to get precise slices on Windows (not sure if `perf` works here, and AMD's tool has its limits)  r/n, so I'll stick to this method.