@lynxlynxlynx 2021-07-09
My test on a SSD and a good (something like 10k RPM) shows slight improvements even for saves with few areas.

However, when testing bg1, after the first resave, the rest took several seconds and I got a few invalid seeks: 
```
[Unknown]: [Streams]: Invalid seek position 2374 in file baldur.sav(limit: 0)
```
Could be from other recent changes perhaps, but in this worst case it slows down saving significantly. Restarting the game and loading that save also takes ages (decompresing the blb: 27661 ms).

Something is fishy — why did the size explode so?
```
$ ls -lh build/gemrb/Cache2/
total 4.0G
-rw-r--r-- 1 kajak kajak  43K Jul  9 16:33 ar3800.are
-rw-r--r-- 1 kajak kajak 4.0G Jul  9 16:33 ares.blb
-rw-r--r-- 1 kajak kajak   20 Jul  9 16:33 default.toh
-rw-r--r-- 1 kajak kajak    0 Jul  9 16:33 default.tot
```