@MarcelHB 2021-07-09
So reproducing for you is: Loading a game, and saving to the same slot, or quicksaving?