@lynxlynxlynx 2021-07-09
It was always manual saving.