@MarcelHB 2021-07-09
Ok, one thing fell victim to my post-wiring of error return codes. The other thing I did not notice came up when resaving the 2nd time, since strlen and the value in zip entry are off by one.