@lynxlynxlynx 2021-07-10
A user reported a crash and it led me back here. Starting a new tob game is a simple way to reproduce. An autosave is scripted and it corrupts memory and boom.

```gdb
#0  0x00000000006b4311 in GemRB::ResourceManager::GetResource (this=0x5028, 
    ResRef=0x1010 <error: Cannot access memory at address 0x1010>, type=327680, silent=true)
    at /home/kajak/games/gemrb/gemrb/core/ResourceManager.cpp:102
#1  0x00000000006b8f25 in GemRB::SaveGame::GetSave (this=0x0)
    at /home/kajak/games/gemrb/gemrb/core/SaveGameIterator.cpp:171
#2  0x00000000006b567c in GemRB::SaveGameAREExtractor::copyRetainedAREs (this=0xa39bb8, 
    destStream=0x7fffffff7240, trackLocations=false)
    at /home/kajak/games/gemrb/gemrb/core/SaveGameAREExtractor.cpp:40
#3  0x0000000000653c42 in GemRB::Interface::CompressSave (this=0xa39780, 
    folder=0x7fffffffb4c0 "/media/kajak/games/bg2-gog/prefix/drive_c/GOG Games/Baldur's Gate 2/save/000000000-Auto-Save", overrideRunning=false) at /home/kajak/games/gemrb/gemrb/core/Interface.cpp:4510
#4  0x00000000006bb176 in GemRB::DoSaveGame (
    Path=0x7fffffffb4c0 "/media/kajak/games/bg2-gog/prefix/drive_c/GOG Games/Baldur's Gate 2/save/000000000-Auto-Save", overrideRunning=false) at /home/kajak/games/gemrb/gemrb/core/SaveGameIterator.cpp:436
#5  0x00000000006ba955 in GemRB::SaveGameIterator::CreateSaveGame (this=0xe23f90, index=0, mqs=false)
    at /home/kajak/games/gemrb/gemrb/core/SaveGameIterator.cpp:635
#6  0x000000000060b41d in GemRB::GameScript::SaveGame (parameters=0x3b972e0)
    at /home/kajak/games/gemrb/gemrb/core/GameScript/Actions.cpp:5921
```