@MarcelHB 2021-07-10
Grml, but easy. Just the only place I forgot checking for a specific nullptr.